добавили переход на станицу игры по клику на кнопку
