+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\my app\\pokemon-game\\src\\App.js":"1","D:\\my app\\pokemon-game\\src\\components\\Footer\\footer.js":"2","D:\\my app\\pokemon-game\\src\\components\\Header\\header.js":"3","D:\\my app\\pokemon-game\\src\\components\\Layout\\Layout.js":"4","D:\\my app\\pokemon-game\\src\\index.js":"5","D:\\my app\\pokemon-game\\src\\components\\PokemonCard\\index.js":"6"},{"size":5936,"mtime":1612000824035,"results":"7","hashOfConfig":"8"},{"size":317,"mtime":1611751703995,"results":"9","hashOfConfig":"8"},{"size":492,"mtime":1611750369926,"results":"10","hashOfConfig":"8"},{"size":866,"mtime":1612244730500,"results":"11","hashOfConfig":"8"},{"size":175,"mtime":1611739956503,"results":"12","hashOfConfig":"8"},{"size":2091,"mtime":1612244509024,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"19463ej",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\my app\\pokemon-game\\src\\App.js",[],"D:\\my app\\pokemon-game\\src\\components\\Footer\\footer.js",[],"D:\\my app\\pokemon-game\\src\\components\\Header\\header.js",[],"D:\\my app\\pokemon-game\\src\\components\\Layout\\Layout.js",[],"D:\\my app\\pokemon-game\\src\\index.js",[],"D:\\my app\\pokemon-game\\src\\components\\PokemonCard\\index.js",[]]
+[{"D:\\my app\\pokemon-game\\src\\App.js":"1","D:\\my app\\pokemon-game\\src\\components\\Footer\\footer.js":"2","D:\\my app\\pokemon-game\\src\\components\\Header\\header.js":"3","D:\\my app\\pokemon-game\\src\\components\\Layout\\Layout.js":"4","D:\\my app\\pokemon-game\\src\\index.js":"5","D:\\my app\\pokemon-game\\src\\components\\PokemonCard\\index.js":"6","D:\\my app\\pokemon-game\\src\\routes\\Home\\home.js":"7","D:\\my app\\pokemon-game\\src\\routes\\Game\\game.js":"8"},{"size":540,"mtime":1612255349819,"results":"9","hashOfConfig":"10"},{"size":317,"mtime":1611751703995,"results":"11","hashOfConfig":"10"},{"size":673,"mtime":1612255284788,"results":"12","hashOfConfig":"10"},{"size":866,"mtime":1612244730500,"results":"13","hashOfConfig":"10"},{"size":175,"mtime":1611739956503,"results":"14","hashOfConfig":"10"},{"size":2078,"mtime":1612246714929,"results":"15","hashOfConfig":"10"},{"size":6190,"mtime":1612255327064,"results":"16","hashOfConfig":"10"},{"size":225,"mtime":1612255137215,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"19463ej",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\my app\\pokemon-game\\src\\App.js",[],"D:\\my app\\pokemon-game\\src\\components\\Footer\\footer.js",[],"D:\\my app\\pokemon-game\\src\\components\\Header\\header.js",[],"D:\\my app\\pokemon-game\\src\\components\\Layout\\Layout.js",[],"D:\\my app\\pokemon-game\\src\\index.js",[],"D:\\my app\\pokemon-game\\src\\components\\PokemonCard\\index.js",[],"D:\\my app\\pokemon-game\\src\\routes\\Home\\home.js",[],"D:\\my app\\pokemon-game\\src\\routes\\Game\\game.js",[]]
