+++ src/App.css
@@ -1,7 +1,1 @@
-.flex{
-    display: flex;
-    justify-items: center;
-    align-items: center;
-    flex-wrap: wrap;
 
-}
