+++ src/routes/Game/game.js
@@ -1,0 +1,12 @@
+
+import React from 'react';
+import s from './style.module.css';
+
+const GamePage=()=>{
+    return(
+        <div className={s.wrapper}>
+            This is GAME PAGE
+        </div>
+    )
+};
+export default GamePage;
