+++ src/routes/Game/style.module.css
@@ -1,0 +1,6 @@
+.wrapper{
+    text-align: center;
+    font-size: 30px;
+    color: #f300b4;
+    padding-top: 40px;
+}
