+++ src/routes/Home/style.module.css
@@ -1,0 +1,7 @@
+.flex{
+    display: flex;
+    justify-items: center;
+    align-items: center;
+    flex-wrap: wrap;
+
+}
